@@ -6,8 +6,8 @@ publish_to: 'none'
 version: 2.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
-  flutter: ">=2.10.5"
+  sdk: ">=2.17.1 <3.0.0"
+  flutter: ">=3.0.1"
 
 dependencies:
 
@@ -29,8 +29,8 @@ dependencies:
   replay_bloc: ^0.2.1
 
   # Providers (riverPod + hooks) for Dependencies Injection and State Management.
-  flutter_hooks: ^0.18.3
-  hooks_riverpod: ^1.0.3
+  flutter_hooks: ^0.18.4
+  hooks_riverpod: ^1.0.4
 
   # login page
   flutter_login: ^3.2.0
@@ -39,12 +39,12 @@ dependencies:
   dio_cookie_manager: ^2.0.0
 
   # Storages
-  hive: ^2.1.0
+  hive: ^2.2.1
   hive_flutter: ^1.1.0
   flutter_secure_storage: ^5.0.2
 
   # router
-  go_router: ^3.0.7
+  go_router: ^3.1.1
 
 dev_dependencies:
   flutter_test:
@@ -53,13 +53,13 @@ dev_dependencies:
   integration_test:
     sdk: flutter
 
-  very_good_analysis: ^2.4.0
+  very_good_analysis: ^3.0.0
 
-  build_runner: ^2.1.10
+  build_runner: ^2.1.11
 
   # App configurations
   flutter_launcher_icons: ^0.9.2
-  flutter_native_splash: ^2.1.6
+  flutter_native_splash: ^2.2.2
 
   # Bloc
   bloc_test: ^9.0.3
@@ -68,7 +68,7 @@ dev_dependencies:
   import_sorter: ^4.6.0
 
   # hive
-  hive_generator: ^1.1.2
+  hive_generator: ^1.1.3
 
 module:
   androidX: true
